Initial import - role to install / enable regular user to use libvirt
Most ansible roles available for libvirt either install libvirt or configures
libvirt with many options and bells and whistles. All you really need for a
simple set up however, is to install libvirt and enable a regular user to use
libvirt.

This role installs libvirt and configures the current running user to
be able to use KVM via libvirt with the typical distro defaults.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>