fix enable-user task on group addition
Without 'append: yes' we remove the user from all the groups it was
originally added to. This is a severe bug. So for instance you may
have a user running this ansible role who was already on the wheel
group, or the sudo group, and after using this role you'd have it
removed from those groups, therefore losing all admin privs.

Fix this for all distros.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>