enable-user: re-order when clause for user check
Without this we'll get a complain that the a variable does not exist,
and it would not as the rule we depend on to check the items we want
to check would not have been formed as they depend on a registration
which would have ran only if a variable was set. So just check for
that variable first as well.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>